# docs/swagger.yaml
basePath: /
host: localhost:3000
info:
  contact: {}
  description: P2P file sharing with content-addressable storage, collection management,
    versioning, merge/fork/pull.
  title: Peerdrive API
  version: "1.0"
paths:
...
  /collections:
    post:
      consumes:
      - application/json
      description: Create a named collection for a user. Duplicate (username, collection_name)
        returns 409.
      parameters:
      - description: Username and collection name
        in: body
        name: body
        required: true
        schema:
          properties:
            collection_name:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: id, username, collection_name
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Duplicate
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new collection
      tags:
      - collections
  /collections/{username}:
    get:
      description: Returns all collections owned by the given username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: data array of collections
          schema:
            additionalProperties: true
            type: object
      summary: List collections for a user
      tags:
      - collections
  /collections/{username}/{collection_name}:
    get:
      description: Returns collection metadata and its path→hash entries
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: collection and entries
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get collection details with entries
      tags:
      - collections
  /collections/{username}/{collection_name}/commit:
    post:
      consumes:
      - application/json
      description: Snapshot all current entries into a version with a commit message.
        Creates a linked version chain.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      - description: Commit message
        in: body
        name: body
        required: true
        schema:
          properties:
            commit_message:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: message and version_number
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Collection not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Commit collection entries as a new version
      tags:
      - collections
  /collections/{username}/{collection_name}/entries:
    post:
      consumes:
      - application/json
      description: Map a path (e.g. "dir/file.txt") to a SHA256 hash within a collection.
        Creates the collection if it doesn't exist.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      - description: Entry path and file hash
        in: body
        name: body
        required: true
        schema:
          properties:
            hash:
              type: string
            path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add an entry to a collection
      tags:
      - collections
  /collections/{username}/{collection_name}/entries/{path}:
    delete:
      description: Delete a path→hash mapping from a collection
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      - description: Entry path (URL-encoded)
        in: path
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Collection not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove an entry from a collection
      tags:
      - collections
  /collections/{username}/{collection_name}/log:
    get:
      description: Returns all committed versions for a collection, newest first
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: data array of versions
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Collection not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get collection version history
      tags:
      - collections
  /collections/{username}/{collection_name}/rollback/{version_id}:
    post:
      description: Restore collection entries from a specific version's snapshot.
        Replaces all current entries.
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Collection name
        in: path
        name: collection_name
        required: true
        type: string
      - description: Version ID to restore
        in: path
        name: version_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid version_id
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Collection not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rollback collection to a previous version
      tags:
      - collections
...
  /sha256sum/{sha256}:
    get:
      description: Download a file using its SHA256 hash as the content identifier.
        Looks up metadata in SQLite, then streams from the appropriate provider (local
        or HTTP).
      parameters:
      - description: 64-character lowercase SHA256 hex string
        in: path
        name: sha256
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "400":
          description: Invalid hash format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download file by SHA256
      tags:
      - download
...
swagger: "2.0"